pyproject.toml
[tool.poetry]
name = "wiigames-categorizer"
version = "0.1.0"
description = "Wii games categorizer for USB Loader GX"
authors = ["imtrisha"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
xmltodict = "^0.14.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
